dominant_color_of_image-main/requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
numpy==1.19.4
olefile==0.46
pandas==1.1.5
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
urllib3==1.26.2
webcolors==1.11.1
Werkzeug==1.0.1
